[Accounting] Handle account balance amount.
For transactions that lack balance amount information, it is replaced with formula to compute from the previous value.

The balance amount in the monthly summary is a formula that uses the balance amount from the previous month.

The initial balance amount is set to 0 as there is no direct way to get the balance information. It can only be derived.